[feat] AvailableTime 엔티티, DTO
