+++ server-quota/src/main/java/com/o1b4/serverquota/entity/AvailableTime.java
@@ -1,6 +1,8 @@
 package com.o1b4.serverquota.entity;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import javax.persistence.*;
@@ -11,12 +13,16 @@ import java.time.LocalTime;
 @Table(name = "availabletime")
 @Getter
 @ToString
+@NoArgsConstructor
+@IdClass(AvailableTimeId.class)
+@AllArgsConstructor
 public class AvailableTime implements Serializable {
-    @Id
-    @ManyToOne
-    @JoinColumn(name="roomid")
-    private ReservationRoom room;
 
+    @Id
+    @Column(name="roomid")
+    private long roomId;
+
+    @Id
     @Column(name = "wday", nullable = false)
     private int wDay;
 
@@ -25,15 +31,5 @@ public class AvailableTime implements Serializable {
 
     @Column(name = "endtime", nullable = false)
     private LocalTime endTime = LocalTime.of(19, 0);
-
-    public AvailableTime() {
-    }
-
-    public AvailableTime(ReservationRoom room, int wDay, LocalTime startTime, LocalTime endTime) {
-        this.room = room;
-        this.wDay = wDay;
-        this.startTime = startTime;
-        this.endTime = endTime;
-    }
 
 }
